+++ doc/source/conf.py
@@ -323,6 +323,11 @@ rst_epilog = """
 .. |VibrationsJob| replace:: :class:`~scm.plams.recipes.vibration.VibrationsJob`
 .. |IRJob| replace:: :class:`~scm.plams.recipes.vibration.IRJob`
 .. |VibrationsResults| replace:: :class:`~scm.plams.recipes.vibration.VibrationsResults`
+.. |ADFFragmentJob| replace:: :class:`~scm.plams.recipes.adffragment.ADFFragmentJob`
+.. |ADFFragmentResults| replace:: :class:`~scm.plams.recipes.adffragment.ADFFragmentResults`
+.. |BANDFragmentJob| replace:: :class:`~scm.plams.recipes.bandfragment.BANDFragmentJob`
+.. |BANDFragmentResults| replace:: :class:`~scm.plams.recipes.bandfragment.BANDFragmentResults`
+.. |NOCVBandFragmentJob| replace:: :class:`~scm.plams.recipes.bandfragment.NOCVBandFragmentJob`
 .. |ADFCOSMORSCompoundJob| replace:: :class:`ADFCOSMORSCompoundJob<scm.plams.recipes.adfcosmorscompound.ADFCOSMORSCompoundJob>`
 .. |ADFCOSMORSConfJob| replace:: :class:`ADFCOSMORSConfJob<scm.plams.recipes.adfcosmorsconformers.ADFCOSMORSConfJob>`